drm/omap: ensure all displays have been probed
DRM offers no ways to add new displays after the DRM driver has been
loaded. This causes issues on boards that have multiple displays, and
omapdrm is loaded when some of them are loaded but not all. The result
is that omapdrm starts with the displays that were loaded at that time,
ignoring the displays loaded later.

This patch changes the behavior so that omapdrm ensures all display are
loaded which have an alias in the .dts file.

The downside is that this prevents omapdrm from loading if, say, the
user has not compiled a kernel module for one of the displays, or
there's some kind of failure in one of the displays. Thus, after this
patch, all the displays have to work, or none does.

TODO before sending to mainline:

The DSS bindings document the aliases as optional, so this won't work in
the general case. The best solution would be to travel the display graph
through the ports and use the components framework, but that would be
more complex to implement.

Signed-off-by: Tomi Valkeinen <tomi.valkeinen@ti.com>