HACK: OMAPDSS: Add panel-tlc59180
DRA7 EVM and DRA72 EVM have a panel for which we don't have proper
drivers. We'd need a driver for the TLC59108 LED chip, which, on those
boards, is used as both GPIO expander and a backlight source. We don't
have such a driver, and so we have this custom DRA7/72 specific panel
driver that does the job.

Signed-off-by: Tomi Valkeinen <tomi.valkeinen@ti.com>