OMAPDSS: HDMI: Do not abort audio playback when display is turned off
Do not abort audio playback when display is turned off. The audio DMA
stops when the display turned off and the audio stream will timeout in
couple of seconds unless the display is enabled again in time. Without
this patch the audio playback is aborted immediately when display is
turned off.

Signed-off-by: Jyri Sarha <jsarha@ti.com>