net/rpmsg: sanity check the input buffer length in sendmsg
The virtio rpmsg transport uses 512 byte buffers, and this is
the maximum data that can be transmitted including the rpmsg
message header size. The individual rpmsg client drivers
therefore can transmit only upto about 496 bytes (based on
16-bytes for rpmsg_hdr), and the rpmsg_sock_sendmsg() uses
a local kernel buffer on the stack for copying the buffer
from userspace and transmitting it.

Perform a sanity check on the length of the buffer passed in
from the userspace in rpmsg_sock_sendmsg() to not exceed the
rpmsg client driver payload size. This fixes a kernel hang
when large messages are attempted to be transmitted.

Reported-by: Robert Tivy <rtivy@ti.com>
Signed-off-by: Suman Anna <s-anna@ti.com>