ARM: dts: am4372.dtsi: disable rfbi
When DSS nodes were added to am4372.dtsi, the rfbi node was not marked
as disabled. This should have been done, as the rule of thumb is to
disable all DSS nodes that are not used, and especially rfbi, as we
don't have a driver for rfbi.

This hasn't caused any problems until the recent addition of omapdss
componentization. Now the omapdss will not load if it cannot probe all
the submodules. This never happens for rfbi, as we don't have a driver,
and thus omapdss never initializes on AM4372.

This patch adds the missing disable.

Signed-off-by: Tomi Valkeinen <tomi.valkeinen@ti.com>