soc: ti: wkup_m3_ipc: Add cpuidle support
Add support for sending a message to the wkup_m3 to enter cpuidle state.
Currently wkup_m3 just places MPU clockdomain in sleep and handles waking
back up.

Also, add a wkup_m3_ping_noirq function internally so that we can send a
message without waiting in a completion as when we send a cpuidle
message we are in noirq context.

Signed-off-by: Dave Gerlach <d-gerlach@ti.com>