ARM: OMAP: DRA7xx: [HACK] Make CAM clock domain SWSUP only
HWSUP on this domian is only working when VIP1 probes.
If only VIP2 on J6 or CAL on J6eco probes the domain does not get enabled.
This might indicates an issue in the HW Auto state-machine for this domain.

Work around in to set the CAM domain to use SWSUP only.

Signed-off-by: Benoit Parrot <bparrot@ti.com>
Signed-off-by: Jyri Sarha <jsarha@ti.com>