ARM: dts: dra7: add the necessary node for vpe support
Add the necessary node and configuration data for the VPE (Video
Processing Engine) hardware block on J6.

The corresponding driver for this entry is in
drivers/media/platform/ti-vpe/vpe.c.

This entry relies on the irq crossbar driver being available, as
it uses the crossbar mapping number and not the irq number.

Signed-off-by: Archit Taneja <archit@ti.com>
Signed-off-by: Darren Etheridge <detheridge@ti.com>
Signed-off-by: Benoit Parrot <bparrot@ti.com>
Signed-off-by: Jyri Sarha <jsarha@ti.com>