media: ti-vpe: vpdma: Convert vpdma into its own helper module
The VPDMA (Video Port DMA) as found in devices such as DRA7xx is
used for both the Video Processing Engine (VPE) and the Video Input
Port (VIP).  The VIP uses a different subset of functionality from
the VPDMA when compared to the VPE.
The purpose of this commit is to upgrade the VPDMA driver to pave
the way for the addition of the VIP driver.  Where possible API's
have not been changed in the VPDMA driver that the VPE was already
using.  However in a couple of cases this was really unavoidable
and the subsequent modifications to the VPE driver are included
in this commit.

Signed-off-by: Benoit Parrot <bparrot@ti.com>
Signed-off-by: Darren Etheridge <detheridge@ti.com>
Signed-off-by: Jyri Sarha <jsarha@ti.com>