spi: davinci: Set prescale value based on register value
Within davinci_spi_get_prescale() the prescale has two
meanings. First one being the calculated prescale value and
then at the end translates it to the prescale value that
will be written to the SPI register.

At first glance this can be confusing especially when
comparing the minimum prescale value against what is seen in
the TRM.

To simplify things make it clear that the calculated
prescale value will always be based on the value that will
be written into the SPI register.

Signed-off-by: Franklin S Cooper Jr <fcooper@ti.com>
Signed-off-by: Sekhar Nori <nsekhar@ti.com>