iommu/omap: Fix debug_read_tlb() to use seq_printf()
[ Upstream commit e203db293863fa15b4b1917d4398fb5bd63c4e88 ]

The debug_read_tlb() uses the sprintf() functions directly on the buffer
allocated by buf = kmalloc(count), without taking into account the size
of the buffer, with the consequence corrupting the heap, depending on
the count requested by the user.

The patch fixes the issue replacing sprintf() by seq_printf().

Signed-off-by: Salva Peiró <speirofr@gmail.com>
Signed-off-by: Joerg Roedel <jroedel@suse.de>
[s-anna@ti.com: cherry-pick commit 'e203db293863' from v4.3]
Signed-off-by: Suman Anna <s-anna@ti.com>