ARM: dts: am437x-gp-evm: Add regulator-always-on and regulator-boot-on for RTC DCDCs
DCDC5 and DCDC6 supply rtc and need to be on for accessing the module.

On A1 revision of the TPS65218, FSEAL bit would be undefined without
coin-cell present which in many cases led to it being set, causing DCDC5
and DCDC6 to stay active and allowing RTC-only mode to work but also
leading to unexplained failures when it was not. On B1 revision, FSEAL
is always 0 when no coin-cell is present so this patch is required on
boards with B1 revision to ever work. This implementation works on
boards with either A1 or B1 revision and makes sure that DCDC5 and DCDC6
always stay active.

Signed-off-by: Dave Gerlach <d-gerlach@ti.com>
[Ported to 4.1]
Signed-off-by: Keerthy <j-keerthy@ti.com>