ARM: AM43xx: Introduce a separate soc_is function for am438x series of SoCs
EPOS evms are fitted with a separate family of am43xx SoCs and are named
am438x series. Adding a separate soc_is function to identify that
particular series of SoCs. This can be done to avoid unnecessarily
registering hwmods like rtc when not needed on EPOS evms.
The last specific number of the part is unknown yet so keeping it generic
am438'x'.

Signed-off-by: Keerthy <j-keerthy@ti.com>