ARM: OMAP2+: pm33xx: Refactor am33xx_push_sram_idle
Separate out the portions of am33xx_push_sram_idle that handles the
generic allocator calls and emif function address lookup from the
actual copy functionailty. We only need to do the allocation during
init as this will be preserved, however we need to redo the copy to the
previously allocated address space after any transition that causes
us to lose SRAM context, like RTC+DDR mode, so that the PM code is
available to us at the same addresses as before for future low power
transitions.

Signed-off-by: Dave Gerlach <d-gerlach@ti.com>