ARM: dts: OMAP4: Update the DSP node
The compatible property for the DSP node is updated to match
the OMAP remoteproc driver needs. The node is moved from the
soc node to the ocp node, as this is better suited to use
with pdata-quirks.

The node is updated with the 'syscon-bootreg', 'iommus' and
'mboxes' properties. The node is disabled for now, and should
be enabled as per the individual product configuration in the
corresponding board dts files.

Signed-off-by: Suman Anna <s-anna@ti.com>