ARM: OMAP2+: Add pdata-quirks for DSP and IPU remoteprocs
The OMAP remoteproc driver performs the device management (reset
control and clocking) for the remote processor sub-systems using
the omap_device API which are limited to only the mach-omap layer.
The OMAP mechanism of using pm_runtime API to achieve this is
insufficient as these devices have hard reset lines which are
managed separately. Use pdata quirks to manage the device reset
and clocking functionality through platform data ops, until the
reset portions are decoupled from omap_hwmod/omap_device into a
separate reset driver.

This patch adds the pdata quirks for both the DSP and IPU processor
subsystems on OMAP4. The pdata quirks do not use a starting address
in the auxdata, as there are only single instances of DSP and IPU
on OMAP4, and this also matches with the current DT node definitions.

Signed-off-by: Suman Anna <s-anna@ti.com>