ARM: dts: am43x-epos-evm: Add pinmux for gpio0 wake
Add pinctrl settings so that gpio0 wake from suspend will be supported
using all buttons on evm. This is accomplished by enabling internal
pull-ups on the row lines of the matrix so that when any button is
depressed the column pins, which are all GPIO0, receive a high signal.

Also, make sure the additional pins do not use internal pullup as they
are connected to external pulldowns and this will cause excess current
drain.

Signed-off-by: Dave Gerlach <d-gerlach@ti.com>