ARM: DRA7: hwmod data: Add MMU data for DSP2
DRA74x has two DSPs, while DRA72x has only one DSP. Each DSP
subsystem has two MMUs, one for the processor core and the
other for the internal EDMA block. Add the hwmod data for
the second DSP on DRA74x only. The MMU associated with the
EDMA port is expected to be mirror-programmed alongside the
primary MMU, with the reset handled by the OMAP IOMMU driver.
The reset data is added to both the MMUs to allow the
omap_hwmod layer to skip the enabling and idling of these
devices, as that would require the reset be released, which
is outside the scope of the hwmod core code. The pdata quirks
will not be added though for the second MMU as the OMAP IOMMU
driver releases the reset once and programs both the MMUs
together.

Signed-off-by: Suman Anna <s-anna@ti.com>