ARM: OMAP2: Add functions to save and restore powerdomain context
The powerdomain control registers are stored in the WKUP powerdomain on
AM33XX, which is lost on RTC-only suspend and also hibernate. This adds
context save and restore functions for those registers.
Sometimes the powerdomain state does not need to change,
perhaps we only need to change memory retention states, so make
sure the restored state is different from the current state before we wait
for a transition.

Signed-off-by: Keerthy <j-keerthy@ti.com>
[Ported to 4.1 and squashed save/restore and conditional restoration]
Signed-off-by: Dave Gerlach <d-gerlach@ti.com>
[Conditional restoration]
Signed-off-by: Russ Dill <Russ.Dill@ti.com>