ARM: dts: dra7-evm: Add pinmux configuration for MMC1 & MMC2
Add pinmux configuration for MMC1 & MMC2 including IODELAY
values suggested in the data manual for the various supported
modes.

Silicon revision 2.0 of DRA74x SoC, available on revision
H and onwards of DRA74x EVM provides for higher MMC/SD
speeds.

This is possible because errata i843 is addressed on this
silicon.

IOdelay data for both silicon revision 1.1 and 2.0 is
maintained. The maximum frequency supported is matches
 that of rev 2.0. pdata quirks must be utilized to
detect presence of rev 1.1/1.0 of silicon and adjust
maximum frequencies as per restrictions documented in i843.

Signed-off-by: Kishon Vijay Abraham I <kishon@ti.com>
Signed-off-by: Sekhar Nori <nsekhar@ti.com>