dmaengine: omap-dma: Support for polled memcpy transfers
When copying small chunks of data the client driver can decide to not use
callbacks and completion (or other means of handling the end of transfer),
but to simply poll the status of the transfer.
This is indicated by the missing DMA_PREP_INTERRUPT flag - so that the
client has no interest in the interrupt.
In such a case the driver will start the transfer without delay and in the
omap_dma_tx_status() callback we will check if the channel is still enabled
or not, if not enabled, DMA_COMPLETED status is returned.

Signed-off-by: Peter Ujfalusi <peter.ujfalusi@ti.com>
Tested-by: Tomi Valkeinen <tomi.valkeinen@ti.com>
Signed-off-by: Jyri Sarha <jsarha@ti.com>