ti_config_fragments: omap5: Change USB Host and SMSC to built in
The omap5 evm has the ethernet connect to the EHCI USB host
and the EHCI USB host is built as module so when doing a
nfs mount the mount fails because the USB host is not
initialized.

Since this configuration is only desired on the omap5
for system test add this config to the system
test config frag.

Signed-off-by: Dan Murphy <dmurphy@ti.com>