remoteproc: fix a compiler warning in debugfs code
The following compiler warning is seen with the Linaro
4.9-based 2015.05 compiler:

  drivers/remoteproc/remoteproc_debugfs.c:
    In function 'rproc_recovery_write':
  drivers/remoteproc/remoteproc_debugfs.c:167:9: warning:
    'buf[4294967295u]' may be used uninitialized in this function
    [-Wmaybe-uninitialized]
  if (buf[count - 1] == '\n')

The above is a false alarm, as the variable 'count' is indeed
checked for an upper limit before using it as an array index.
In anycase, re-work the logic and fix the compiler warning by
introducing another variable. There is no change in functionality
of the debugfs behavior.

Signed-off-by: Suman Anna <s-anna@ti.com>