dmaengine: edma: HACK: Disable memcpy on DRA7xx family
commit a41573a74a06c517cce2e5d6adb2b0d9bb9d4976 ti-linux-4.1.y

DRA7xx family of SoCs have DMA crossbar on front of the eDMA (like with
sDMA). Due to the legacy in the eDMA driver stack when the crossbar is in
use we can not support MEMCPY operation via eDMA since it can breakm the
HW triggered channels.
The MEMCPY can be done via the sDMA.

Signed-off-by: Peter Ujfalusi <peter.ujfalusi@ti.com>
Signed-off-by: Gabriel Huau <ghuau@ti.com>