HACK: drm/omap: Add omapdrm plugin API
This patch enables SGX driver to be added as a plugin to omapdrm. This
is a HACK, and must be reverted/rewritten when proper DRM-SGX interop is
implemented.

As this is a HACK, the code is wrapped with CONFIG_DRM_OMAP_SGX_PLUGIN,
and grouped together and separated from the main omapdrm code when
possible. This means checkpatch will complain a bit, but this should
make it easier to revert this later and to avoid conflicts with
mainline.

Main changes involved:

1. SGX specific GEM VM operations
	SGX requires contiguous memory for both texture memory as well
	as framebuffers.

	Memory allocation of FB is done through omapdrm and is guaranteed
	to be contiguous.

	Texture memory can come from:
	a. user space allocated memory
	b. memory is allocated by other cores
	c. memory comes from CMA

	Texture memory can be non-contiguous, but is wrapped as GEM
	objects which enables the use of TILER to map them as contiguous
	memory to the SGX HW.

2. Support for ioctls from plugin driver
	SGX driver registers as a plugin to the omapdrm driver. During
	registration, SGX specific ioctls are added to omapdrm. This allows
	user space to control specific SGX feature sets using the DRM FD.

3. Export GEM functions
	SGX driver needs to work directly on GEM objects for DSS
	synchronization, getting TILER address, etc.

Signed-off-by: Rob Clark <rob@ti.com>
Signed-off-by: Subhajit Paul <subhajit_paul@ti.com>
Signed-off-by: Tomi Valkeinen <tomi.valkeinen@ti.com>
Signed-off-by: Anand Balagopalakrishnan <anandb@ti.com>
Signed-off-by: Jyri Sarha <jsarha@ti.com>