ARM: dts: DRA7: Add the PRU-ICSS nodes
Add the DT nodes for the PRU-ICSS1 and PRU-ICSS2 processor
subsystems that are present on AM57xx family of SoCs. The
two PRU-ICSSs are identical to each other. They are not
supported on DRA74x and DRA72x SoCs in general, so the
nodes are added in disabled state to the common dra7 DTS
file. They should be enabled only in the AM57xx related
board files.

The DT nodes uses all standard properties. The regs property
defines the addresses for various modules within the PRUSS,
including all the Data RAMs, the Instruction RAM per core,
the INTC, CFG, Debug and Control sub-modules. The two PRU
cores within the PRUSS are represented as child nodes (so
platform devices themselves) of the PRUSS parent node.

The PRU-ICSSs on AM57xx are very similar to the PRUSS in
AM33xx and AM437x except for variations in the RAM sizes
and the number of interrupts coming into the MPU INTC. The
interrupt events into the PRU-ICSS also requires programming
of the corresponding crossbars properly.

Signed-off-by: Suman Anna <s-anna@ti.com>