arm: dtsi: dra7: Add port B endpoints in VIP nodes
Each VIP slice has a 24bit port A and an 8bit port B for video capture.
For each of this port, there has to be an endpoint device node which
should connect to the camera endpoint node.

Add the endpoint device nodes for port B. The reg property of the endpoint
node describes the slice and port it belongs to.

In addition each port node now support the 'status' node, so they can be
individually enabled. Default port status is disabled.

Signed-off-by: Nikhil Devshatwar <nikhil.nd@ti.com>
Signed-off-by: Benoit Parrot <bparrot@ti.com>
Signed-off-by: Jyri Sarha <jsarha@ti.com>