drm/omap: fix handling of the crtc properties
We have some properties which are crtc specific, and then we have
properties for planes. The properties for planes are also used for
crtcs.

In omap_crtc.c's omap_crtc_atomic_set_property and
omap_crtc_atomic_set_property we check if the given property is a plane
property, and if so, call the relevant function in the plane object
instead. This check is done by comparing the property to the plane
properties which have been added by omapdrm.

This doesn't work correctly, as there are common properties for planes,
which are added by the DRM framework. Thus common plane properties such
as src_x, rotation, etc. fail our check, and they are handled as crtc
properties, leading to set/get_property returning an error. Which then
breaks the userspace.

This patch fixes the issue by changing the code to check if the given
property is a crtc property. This works much better, as when
omap_crtc_atomic_set/get_property is called, the caller (DRM framework)
has already handled common crtc properties. Thus
omap_crtc_atomic_set/get_property can only check against omapdrm's
private crtc properties.

If that check shows it is not a crtc property,
omap_crtc_atomic_set/get_property routes the call back to DRM framework,
for the plane object. Which then will first handle common plane
properties, and call omap_plane_atomic_set/get_property if necessary.

Signed-off-by: Tomi Valkeinen <tomi.valkeinen@ti.com>
Reported-by: Robert Nelson <robertcnelson@gmail.com>
Tested-by: Robert Nelson <robertcnelson@gmail.com>
Signed-off-by: Jyri Sarha <jsarha@ti.com>